image-action/requirements.txt
ImageHash==4.1.0
PyGithub==1.51
Pillow==7.2.0